requirements.txt
azure-functions==1.20.0
azure-cosmos==4.6.0        # <─ add this line (or latest version)
azure-eventhub
requests
PyJWT
